Fixed namespaced and moved back wrong class
