+++ src/Umbraco.Infrastructure/PropertyEditors/RichTextEditorBlockValidator.cs
