+++ src/Umbraco.Infrastructure/PropertyEditors/Validators/IRichTextRequiredValidator.cs
@@ -1,6 +1,6 @@
 using Umbraco.Cms.Core.PropertyEditors;
 
-namespace Umbraco.Cms.Infrastructure.PropertyEditors.Validators;
+namespace Umbraco.Cms.Core.PropertyEditors.Validators;
 
 internal interface IRichTextRequiredValidator : IValueRequiredValidator
 {
